refactor(core/plugins): move colorscheme presets
