+++ lua/default.lua
@@ -5,9 +5,9 @@ local f = io.open(colorscheme_cache, "r")
 if f ~= nil then
     local colorscheme = f:read "*a"
     f:close()
-    vim.g.user_colorscheme = require("core.colorscheme")[colorscheme]
+    Ice.colorscheme = require("plugins.colorscheme")[colorscheme]
 else
-    vim.g.user_colorscheme = require("core.colorscheme")["tokyonight"]
+    Ice.colorscheme = require("plugins.colorscheme")["tokyonight"]
 end
 
 default.keymap = require "core.keymap"
